Cleaning up CudiBundle

+++ module/CudiBundle/Entity/Sale/Article.php
@@ -290,7 +290,7 @@ class Article
      */
     public function setPurchasePrice($purchasePrice)
     {
-        $this->purchasePrice = round(str_replace(',', '.', $purchasePrice) * 100);
+        $this->purchasePrice = (int) round(str_replace(',', '.', $purchasePrice) * 100);
 
         return $this;
     }
@@ -310,7 +310,7 @@ class Article
      */
     public function setSellPrice($sellPrice)
     {
-        $this->sellPrice = round(str_replace(',', '.', $sellPrice) * 100);
+        $this->sellPrice = (int) round(str_replace(',', '.', $sellPrice) * 100);
 
         return $this;
     }
